Improve handling of initial wizard values
